# public/debug-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Imo State Ministry of Works API",
    "version": "1.0.0",
    "description": "API documentation for the Imo State Ministry of Works"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://imostateministryofworks.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/google/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Handle Google OAuth callback",
        "security": [],
        "responses": {
          "200": {
            "description": "User authenticated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "1"
                        },
                        "username": {
                          "type": "string",
                          "example": "johndoe"
                        },
                        "email": {
                          "type": "string",
                          "example": "johndoe@example.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "OAuth authentication failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get all projects",
        "security": [],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "1"
                          },
                          "title": {
                            "type": "string",
                            "example": "Road Construction"
                          },
                          "description": {
                            "type": "string",
                            "example": "Highway expansion project"
                          },
                          "imageUrl": {
                            "type": "string",
                            "example": "https://example.com/image.jpg"
                          },
                          "cloudinaryId": {
                            "type": "string",
                            "example": "cloudinary_id_123"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "planned",
                              "in-progress",
                              "completed"
                            ],
                            "example": "planned"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-07-20T18:58:00Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-07-20T18:58:00Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error fetching projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create a new project with image upload",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Project data and image file",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Road Construction"
                  },
                  "description": {
                    "type": "string",
                    "example": "Highway expansion project"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "planned",
                      "in-progress",
                      "completed"
                    ],
                    "example": "planned"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "The project image to upload"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "1"
                        },
                        "title": {
                          "type": "string",
                          "example": "Road Construction"
                        },
                        "description": {
                          "type": "string",
                          "example": "Highway expansion project"
                        },
                        "imageUrl": {
                          "type": "string",
                          "example": "https://example.com/image.jpg"
                        },
                        "cloudinaryId": {
                          "type": "string",
                          "example": "cloudinary_id_123"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "planned",
                            "in-progress",
                            "completed"
                          ],
                          "example": "planned"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-07-20T18:58:00Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-07-20T18:58:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error creating project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Image upload failed"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Invalid or missing token"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}